pyproject.toml
[tool.poetry]
name = "battery-simulator"
version = "1.0.0"
description = "A Python-based battery test data simulator that generates realistic lithium-ion battery cycling data"
authors = ["Battery Analytics Team"]
license = "MIT"
readme = "README.md"
packages = [{include = "battery_simulator"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.21.0"
pandas = "^2.0.0"
scipy = "^1.7.0"
pyyaml = "^6.0"
matplotlib = "^3.4.0"
click = "^8.0.0"
tqdm = "^4.62.0"
pydantic = "^2.0.0"
plotly = "^5.18.0"

# Optional dependencies for web app (requires Python 3.10+)
streamlit = {version = "^1.28.0", optional = true, python = ">=3.10"}

[tool.poetry.extras]
webapp = ["streamlit"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
battery-simulator = "battery_simulator.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

